[BP] Organising Rules into smaller context specific chunks
