+++ csv-validator-core/src/main/scala/uk/gov/tna/dri/schema/package.scala
@@ -14,7 +14,7 @@ package object schema {
   val XsdDateRegex = "[0-9]{4}-[0-9]{2}-[0-9]{2}"
   val UkDateRegex = "(((0[1-9]|(1|2)[0-9]|3[0-1])\\/(0(1|3|5|7|8)|1(0|2)))|((0[1-9]|(1|2)[0-9]|30)\\/(0(4|6|9)|11))|((0[1-9]|(1|2)[0-9])\\/02))\\/[0-9]{4}"
   val XsdTimeRegex = "([0-1][0-9]|2[0-4]):(0[0-9]|[1-5][0-9]):(0[0-9]|[1-5][0-9])(\\.[0-999])?((\\+|-)(0[1-9]|1[0-9]|2[0-4]):(0[0-9]|[1-5][0-9])|Z)?"
-  val PositiveIntegerRegex = "[0-9]+"
-  val UkDateFormat = "dd/MM/YYYY"
   val PartUkDateRegex = """(([0\?][1-9\?])|([1-2\?][0-9\?])|([3\?][0-1\?])|\*)\/(January|February|March|April|May|June|July|August|September|October|November|December|\?|\*)\/([0-9\?]{4}|\*)"""
+  val UkDateFormat = "dd/MM/YYYY"
+  val PositiveIntegerRegex = "[0-9]+"
 }
